pyproject.toml
[tool.poetry]
name = "tm-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
celery = "^5.4.0"
redis = "^5.1.1"
jinja2 = "^3.1.4"
python-telegram-bot = "^21.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
